Use validations on order workshop and accomodation

+++ api/rest/orders.py
@@ -4,12 +4,13 @@ from django.db.utils import DatabaseError
 from django.shortcuts import get_object_or_404
 
 from rest_framework import permissions, response, serializers, status, viewsets
+from rest_framework.serializers import ValidationError
 
 from .payments import PaymentSerializer
 from .reservations import ReservationSerializer
 
 from ..models import Food, Meal, MealReservation, Order, Reservation, Soup,\
-    ParticipantStay, Workshop, Year
+    ParticipantStay, Workshop, Year, Accomodation
 
 
 class OrderSerializer(serializers.ModelSerializer):
@@ -44,35 +45,80 @@ class OrderSerializer(serializers.ModelSerializer):
 
 
 class CreateOrderSerializer(serializers.Serializer):
-    workshop = serializers.IntegerField(required=False, allow_null=True, default=None)
+    workshop = serializers.IntegerField(
+        required=False,
+        allow_null=True,
+        default=None,
+    )
     meals = serializers.ListField()
     stayLength = serializers.ListField(required=False)
     accomodation = serializers.IntegerField()
     accomodationInfo = serializers.BooleanField(required=False)
+    year = None
+
+    def __init__(self, *args, **kwargs):
+        try:
+            self.year = Year.objects.filter(current=True).order_by('-year').first()
+        except Year.DoesNotExist:
+            self.year = None
+        serializers.Serializer.__init__(self, *args, **kwargs)
+
+    def validate(self, data):
+        if not self.year:
+            raise ValidationError('orders.noYearActive')
+        openOrders = Order.objects.filter(
+            canceled=False,
+            participant=self.user.participant,
+            year=self.year,
+        ).count()
+        if openOrders > 0:
+            raise ValidationError('orders.alreadyHaveOne')
+
+        return data
+
+    def validate_accomodation(self, value):
+        if not value:
+            return
+        accomodation = None
+        try:
+            accomodation = Accomodation.objects.get(id=value)
+        except Accomodation.DoesNotExist:
+            raise ValidationError('orders.invalidSelection')
+        if accomodation.year.id != self.year.id:
+            raise ValidationError('orders.invalidSelection')
+        if not accomodation.has_free_capacity():
+            raise ValidationError('orders.noCapacity')
+        return accomodation
+
+    def validate_workshop(self, value):
+        if not value:
+            return
+        workshop = None
+        try:
+            workshop = Workshop.objects.get(id=value)
+        except Workshop.DoesNotExist:
+            raise ValidationError('orders.invalidSelection')
+        if workshop.year.id != self.year.id:
+            raise ValidationError('orders.invalidSelection')
+        if not workshop.has_free_capacity():
+            raise ValidationError('orders.noCapacity')
+        return workshop
 
     def create(self, validated_data): # noqa
-        try:
-            year = Year.objects.filter(current=True).order_by('-year').first()
-        except ObjectDoesNotExist:
-            year = None
-
-        if not year:
-            return None
-
         total_price = 0
 
-        if validated_data['stayLength']:
-            price_level = year.get_actual_price_level()
+        if 'stayLength' in validated_data and validated_data['stayLength']:
+            price_level = self.year.get_actual_price_level()
             currentStay = ParticipantStay.objects.filter(
                 participant=self.user.participant,
-                year=year
+                year=self.year
             )
             for stay in currentStay:
                 stay.delete()
             for day in validated_data['stayLength']:
                 stay = ParticipantStay(
                     participant=self.user.participant,
-                    year=year,
+                    year=self.year,
                     date=day,
                 )
                 stay.save()
@@ -80,9 +126,9 @@ class CreateOrderSerializer(serializers.Serializer):
             if price_level and not validated_data['workshop']:
                 total_price += len(validated_data['stayLength']) * price_level.entryFee
 
-        if 'workshop' in validated_data and validated_data['workshop']:
-            workshop = Workshop.objects.get(id=validated_data['workshop'])
-            workshop_price = workshop.get_actual_workshop_price(year)
+        if validated_data['workshop']:
+            workshop = validated_data['workshop']
+            workshop_price = workshop.get_actual_workshop_price(self.year)
             total_price += workshop_price.price
         else:
             workshop = None
@@ -97,10 +143,10 @@ class CreateOrderSerializer(serializers.Serializer):
             participant=self.user.participant,
             price=total_price,
             accomodation_info=validated_data.get('accomodationInfo', False),
-            year=year,
+            year=self.year,
         )
         reservation = Reservation.objects.create(
-            accomodation_id=validated_data['accomodation'],
+            accomodation=validated_data['accomodation'],
             workshop_price=workshop_price,
             order=order,
         )
@@ -148,26 +194,16 @@ class OrderViewSet(viewsets.ModelViewSet):
         return super().retrieve(request, *args, **kwargs)
 
     def create(self, request):
-        try:
-            year = Year.objects.filter(current=True).order_by('-year').first()
-        except ObjectDoesNotExist:
-            year = None
         serializer = CreateOrderSerializer(data=request.data)
         serializer.user = request.user.participant
-        if year and serializer.user.participant and serializer.is_valid():
-            openOrders = Order.objects.filter(
-                canceled=False,
-                participant=request.user.participant,
-                year=year,
-            ).count()
-            if openOrders == 0:
-                serializer.save()
-                return response.Response(
-                    OrderSerializer(
-                        instance=serializer.instance,
-                        context={'request': request},
-                    ).data,
-                )
+        if serializer.user.participant and serializer.is_valid():
+            serializer.save()
+            return response.Response(
+                OrderSerializer(
+                    instance=serializer.instance,
+                    context={'request': request},
+                ).data,
+            )
         return response.Response(
             serializer.errors,
             status=status.HTTP_400_BAD_REQUEST,
